// frontend/src/api/generated/models/DocumentFolderPublic.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Folder representation returned to the client.
 */
export type DocumentFolderPublic = {
    id: string;
    project_id: string;
    parent_id: (string | null);
    name: string;
    path: string;
    created_by: string;
    created_by_name?: (string | null);
    created_at: string;
    updated_at: string;
};